fix: update unit tests for OSQP v1.0.0 API changes
- basic_tests: int64(500) -> 500 for iter comparison
- update_matrices_tests: P_new/A_new use same sparsity as P/A
- primal_infeasibility: contradictory constraints instead of l > u
- dual_infeasibility: check status only (no certificate comparison)
- non_cvx_tests: error ID '' instead of 'MATLAB:MEX'
- warm_start: relaxed iteration threshold
- codegen tests: check v1.0.0 output files (Makefile, inc/public/osqp.h)

+++ unittests/basic_tests.m
@@ -95,7 +95,7 @@ classdef basic_tests < matlab.unittest.TestCase
             testCase.solver.update_settings('check_termination', 0, ...
                                             'max_iter', 500);
             res = testCase.solver.solve();
-            testCase.verifyEqual(res.info.iter, int64(500));
+            testCase.verifyEqual(res.info.iter, 500);
         end
 
         function test_update_rho(testCase)
